Refactored API to support Read/Write
By removing the footer and only using a header, we can skip needing to
validate the entire payload. This does mean that truly random bytes
being upgraded to being dataversioned are technically able to be
detected as dataversioned, assuming they start with the magic code and
pass muster as an ordered-varint.

But, the use case of this library is adding format transitions and
version transition support across any serialization type. Since
serialization types are unlikely to start with our magic code, it's a
safer assumption than we can make when we assume the input is truly
random data.