Added basic format test
I noticed in coverage that technically the serialize() implementation
was not getting tested fully. The default implementation is sound, and
is a good minimal approach, but it needs to be overridden by any formats
that check lengths before allocating, which Bincode supports.

However, I still liked the idea of serialize() offering this capability
by default, so I just wrote a quick little format that only implements
the missing methods. Since it returns a value from serialized_size, it
will follow the other path which most likely other crates will not use
-- but who knows.