Doc links in READMEs now work "better"
This is the practical application of the new features in RustMe. I grew
tired of how this approach to readmes meant that types linked to from
the README needed to be full URLs. This new system allows customizing
output when rendering a file that has been noted as `for_docs`. We use
this feature to render intra-docs links when outputting the
.crate-docs.md files.

This means that the crate docs will always be links to the same version
of the documentation you're browsing, while the README files will point
to whatever public URL you wish to.